Moved Methods generateHash, generateAddress and VerifyAddress to Repository in infrastructure
